+++ app/models/transfer/task_list.rb
@@ -7,7 +7,8 @@ class Transfer::TaskList < ::BaseTaskList
           Transfer::Task::HandoverTaskForm,
           Transfer::Task::StakeholderKickOffTaskForm,
           Transfer::Task::MainContactTaskForm,
-          Transfer::Task::RequestNewUrnAndRecordTaskForm
+          Transfer::Task::RequestNewUrnAndRecordTaskForm,
+          Transfer::Task::CheckAndConfirmFinancialInformationTaskForm
         ]
       },
       {